pyproject.toml
[tool.poetry]
name = "guido"
version = "0.1.0"
description = ""
authors = ["Nils Freyer <freyer@fh-aachen.de>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
click = "^8.1.3"
tqdm = "^4.64.1"
spacy = "^3.4.2"
pm4py = "^2.2.32"
gensim = "^4.2.0"
omegaconf = "^2.3.0"
hydra-core = "^1.3.1"
transformers = "^4.25.1"
nltk = "^3.8.1"
mlflow = "^2.1.1"



[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"